V0.1, Fix some bug and build a new class
(1) A new class named "VotingScore" had been built in bowVocabulary.h,
which is used to caculate the score of each voting point and display
them.
(2) Replace the old function "macth_a_new_image" with a new heavy load
one.
(3) After this vision, if you change the "sample_size" paramenter in the
config file, DO NOT forget modify them in bowVocabulary.cpp too. It's in
the macro part.
(4) I rewrite the code of voting part, so there are less but rubust
voting points left.:D